Make SamenessKeys copyable.
The samenessKey field of MsgPlaceholderNode is used to decide if 2 placeholders are logically equivalent.  As such it often stores references to the AST.  When copying the AST this means new copies might end up with references to old trees which can lead to surprising results.

This fixes the bug by adding copy constructors and relying on the reflistener system so that the sameness keys can get updated when copies are ocurring.

This did expose one interesting issue in how `genders=` messages are handled.  In this scenario we intentionally copy parts of the tree in order to create a select call.  e.g.

{msg desc=... genders="$userGender"}
  Hello {call .formatName phname="Name" /}
{/msg}

gets rewritten to

{msg desc=... genders="$userGender"}
  {select $userGender}
    {case 'female'}
      Hello {call .formatName phname="Name" /}
    {case 'male'}
      Hello {call .formatName phname="Name" /}
    {case 'other'}
      Hello {call .formatName phname="Name" /}
   {/select}
{/msg}

however, all the placeholders should be treated identically by the translation system.  To fix this we manually copy the sameness keys of all the forked placeholders to point at a single instance.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=317388552