Add support for VE metadata proto extensions to [].
This adds a parameter to [] to pass in the protos used by metadata extensions in the logging config. This also adds an "annotated" logging config, that contains additional details about where the VE metadata is located. This generates the "annotated" logging config as part of the [] and updates the compiler to consume the annotated logging config.

Later CLs will add support for accessing the metadata in Soy.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=317925947