Fix bind() calls in default values in jbcsrc by allowing expression detachment. I believe this is logically consistent with handling of other default values in Soy -- functions are allowed, as long as they are 'pure'. I believe bind() is pure, as long as its children are constant/pure.
(Alternatively, we could modify SoyTreeUtils.getNonConstantChildren() to treat bind() as non-constant, to get the correct error message, but I thought this would be better. Also note that at a cursory reading, there seems to be some inconsistency between the handling of data access nodes in SoyTreeUtils.getNonConstantChildren() and ExpressionCompiler.CanCompileToConstantVisitor, but I'm not sure if this actually matters in practice outside of this case).

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=321167688