Roll forward of []Load VE metadata on the client.
This generates code to reference the VE metadata in the generated metadata file and puts the metadata in the VisualElement. A later Cl will expose this metadata via a method.

NEW: Updated an additional reference to the $$VisualElement constructor.

Automated [] rollback of changelist 323036011.

*** Reason for rollback ***

Roll forward, ensuring all references to the $$VisualElement constructor are updated. One additional reference is updated in this CL. [][]and []updated the remaining references.

*** Original change description ***

Automated [] rollback of changelist 323021622.

*** Reason for rollback ***

There are references to JS $$VisualElement constructor without the additional parameter:

[](fails)
[](fails)

*** Original change description ***

Load VE metadata on the client.

This generates code to reference the VE metadata in the generated metadata file and puts the metadata in the VisualElement. A later Cl will ex...

***

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=323613492