Use invoke dynamic for template calls.
In order to support the complex classloading logic that [] does we added a heuristic in the compiler to call through RenderContext.getTemplateFactory for every template call that crosses the library level.  This supports a complex classloader setup where every compilation unit is potentially owned by a different classloader.  The problem is that even in the happy case we need to call through RenderContext.  The cost of this is small (a few method calls, hash map lookup, call Compiledtemplate.Factory.create()), but still significant compared the the alternate (call 'new foo()')

The solution is to delay the binding decision until runtime, this is the exact usecase for the jvm bytecode instruction 'invokedynamic'.  What this does is allow the runtime to decide how to bind methods to callsites.  So instead we can define a bootstrap method that simply asks what the classloading relationship is between 2 templates.  If they are visible from each other, we can generate a direct call, if not we can go through the slowpath of renderContext.

This theoretically allows for even fancier optimizations for the slowpath where we could bind to the actual factory and then reset when the classloader is recycled, but this would require new APIs for resetting callsites so this work is deferred.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=325280921