Ensure that CSP nonce values inserted by the compiler are base64 encoded.
The Soy compiler adds `nonce` attributes to certain HTML tags to allow applications to use ContentSecurityPolicy directives with nonce values.  This involves rewriting `<script>` tags and certain `<link>` tags to add the `nonce`.

For example

`<script>...</script>`

 becomes

`<script {if $csp_nonce}nonce="{$csp_nonce}"{/if}>...<script>`

This ensures that all scripts visible to the Soy compiler are automatically allowed by the ContentSecurityPolicy.

Currently, the Soy compiler does no special escaping of the `$csp_nonce` value, it simply applies the normal behavior for HTML attributes.

This CL imposes the additional requirement that the value is a `base64` string, this is a defense in depth approach for applications in which the `csp_nonce` value is attacker controlled.

This should not be a breaking change for applications, any application in which this filter would reject the nonce value should be already broken as their nonces won't match.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=328197206