Set Logger on SoyElement when constructing it.
[]
We found that the cause was that the Logger was null when we used SoyElements to re-render an element, but worked otherwise.

I tracked it to here, and found that the Logger is not passed along to the SoyElement when it is constructed, causing issues.

[]

The SoyElement then smoothly passes this along to the new IdomRenderer when we call ".render()"

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=328371344