Delete the error type, use the unknown type instead.
This simplifies various parts of the compiler and ensures that the `$error$` string never shows up in error messages.

Given that an error was already reported, it is hard to see how this would go wrong.  This is motivated by an attempt to simplify our assignability rules for the unknown type, it seemed that the ErrorType should have similar behavior, but at that point, why do we need it at all?

It may be useful to have a separate value to represent a failure of some type operation.  However, i think `Optional<SoyType>` does a much better job and we are already using that in some cases.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=330549005