Split SoyType.isAssignableFrom into 2 methods isAssignableFromStrict and isAssignableLoose
The former has the same semantics as the current isAssignableFrom method.  isAssignableFromLoose is new and it allows for the unknown type to be assigned to everything.

Most calls to isAssignableFrom are simply migrated to isAssignableFromStrict which is fully compatible, a few which were already special casing the unknown type at the callsite are simplified to use isAssignableFromLoose.

Switching to isAssignableFromLoose is a change in behavior since it not only allows values of type `?'` to be used but also more complex things like `[foo: ?]` becomes assignable to `[foo: string]`.

Motivation:  The soy compiler currently has a number of weird rules around the unknown type.  Logically the unknown type should work as a 'bottom type' (https://en.wikipedia.org/wiki/Bottom_type), but our core assignability rules don't treat it as such.  Instead it is special cased in a number of places.  This cl seeks to standardize the two main special cases, and regularize the handling of the unknown type when mixed with complex composite types (e.g. map,record,list).  Future CLs will use this to simplify callsite type checking.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=331857538