Simply the handling of the unknown type when type checking calls
Currently the unknown type is special cased when passing parameters.  A value typed as ? can be passed to most strictly typed parameters.  However, it doesn't apply to more complex types.  For example, `map<string, ?>` cannot be assigned to `map<string, SomeProto>` even though `?` can be assigned to `SomeProto`.

Because Soy templates employ runtime type checking for their parameters, this will allow soy template calls to act like runtime type casts.  Our runtime type checking isn't perfect but this assignability behavior appears to be consistent with the 'any' type in TS.

The benefit of this is that users who have loosely typed values can improve their typing by extracting templates.  This is useful if the proper types can't be 'moved up', for example if a plugin is written in a generic way it may be impossible to fully specify the return type.

This does drop a number of special cases in the compiler that had moderately well justified intentions.  Notably that loosely typed values couldn't be passed to parameters of type `map`, `ve`, or `ve_data`.  As well as a check that protos couldn't be passed to parameters of type `any`. It is possible to preserve these but their existence feels arbitrary and they don't appear to be protecting us against any particularly unique failure modes.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=332036225