Drop the paramsToRuntimeCheck optimization
This is a Tofu only runtime optimization that allows tofu to skip runtime type checks when the compiler has already proven that the types are compatible.  This was important to preserve lazy evaluation (by skipping the type check, we delay resolving the value which was important for futures).

This optimization was added by me in cl/79881671 (November 2014) see full documentation here: []

In the intervening years we have devised a better solution for this, SoySauce which supports this feature natively, albeit in a different way.  Users who rely on this optimization for streaming have mostly been migrated to SoySauce so this is unlikely to be important.

Also, notably, this is not behavior preserving.  Some templates may now fail with type checking errors where they wouldn't before (for example, if the parameter isn't used on the given code path).

The motivation is weak, mostly this change is intended to simplify the CheckTemplateCallsPass.  Supporting tofu only optimizations in the AST no longer seems to be a reasonable tradeoff.

GITHUB_BREAKING_CHANGES=some tofu templates will now report runtime type checking errors more eagerly, some lazy evaluation in tofu may no longer work as expected due to more eager type checking.  For improved lazy rendering performance it is suggested that you migration to SoySauce.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=332870920