Fix a broken condition in the LazyClosureCompiler
We have an optimization where a trivial let or param can be compiled to a constant. However, the same code path is used for compiling message placeholders.  Message placeholders occasionally need to generate wrapping calls to integrate with VE logging. for example:

{msg desc="..."}
  {velog}<div>...</div>{/velog}
{/msg}

because `<div>` is turned into a placeholder and is also attached to a logging command when generating the closure for it we attach an 'extra code compiler', in this scenario we cannot optimize to a constant.  However, we got our conditions backwards which means we basically never hit this optimization.

There is actually a bug here but it is extremely unlikely to be hit given current usages.  If there was a `velog` around a void element, we would promote to a constant and drop the logging commands!  Luckily this never happens in practice.

[]

This should also improve runtime due to work being moved into static initializers.  []

[]

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=341895810