Fix a bug in how streaming print directives are constructed.
Streaming print directives operate as a set of wrapping LoggingAdvisingAppendable instances.  This allows us to continue streaming bytes even while ensuring that values are escaped.  This is important for performance to ensure we don't buffer large amounts of data simply to pass it through escapers and also to preserve our logging semantics.

Some escapers have relatively trivial semantics, e.g. `|escapeJsString` there are simple per-character replacement algorithms.  But others e.g. `|cleanHtml` need to look at multiple characters at a time in order to make decisions, for this they use internal buffers, to ensure these buffers don't swallow characters it is necessary to flush them when the appendable is no longer in use.

Previously we used a strategy of giving such appendables a 'close' method and informing the compiler that it would need to be called.  The compiler would then enssure that all close methods were called in the correct order.  For example, given

`$a|e1|e2|e3` if `|e1|` and `|e3` are closeable we would generate code like

```
tmp1 = output;
tmp2 = createE3(appendable);
tmp3 = createE3(tmp2)
tmp4 = createE1(tmp3);
curren[]pendable = tmp4;
appendable = JbcsrcRuntime.propgateClose(curren[]pendable, ImmutableList.of(tmp4, tmp2));
... use curren[]pendable...

curren[]pendable.close();

```

This worked correctly, but the use of the temporary variables made the logic difficult to follow and hid a bug.  The bug occurred when there were no closeable appendables or the last wrapper was closeable.  In these cases we can skip the call to `propogateClose`.  But due to a bug we ended up generating code like:

```
tmp = createWrapper()
curren[]pendable = createWrapper();
... use curren[]pendable ...

curren[]pendable.close()
```

Notice this dead store to an unused temporary.

The fix is to simplify the way these close calls work and move it from the compiler to the runtime library as much as possible.  For that I have added a new method `flushBuffers(int depth)` to `LoggingAdvisingAppendable`.

This method is defined to flush the buffers of these intermediate appendables and propagate the call to delegates as long as depth is `> 0`.

This results in a slight runtime performance improvement: []

as well as a slight gencode size improvement:

[]

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=342887081