Fix lazy resolution by adding a "DetachableSoyValueProviderProvider" that resolves to a SoyValueProvider and calls renderAndResolve.
This is beneficial to preserve lazy evaluation and streaming when dealing with certain expressions that may contain non trivial values e.g. `$b ? $bar : $baz`.  we would like to be able to stream the values in $bar or $baz, however the current design of DetachableSoyValueProvider doesn't allow for that.  To preserve this kind of laziness we need a new baseclass which this CL provides.

[]

GITHUB_BREAKING_CHANGES=NA

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=345321125