Remove the `detachForCall` special case, replace the use of `detachForCall` with `detachForRender`
The reason for the difference was based on the theory that if a `call` command was going to trigger a detach operation once, it would be likely to do it multiple times, so we could save time saving and restoring locals by only doing it when the call completed.

In other words, we would generate code that looks like:

```java
...
case CONTINUE:
  goto CONTINUE;
...
RenderResult result = template.render(out, context);
if (!result.isDone()) {
  .. save locals..
  state = CONTINUE
  return result;
} else {
  goto DONE;
}
CONTINUE:
result = template.render(out, context);
if (!result.isDone()) {
  return result;
}
.. restore locals...
DONE:
```

Instead we will now generate the same logic we do for print nodes:

```java
...
case BEGIN:
  ..restore locals:
  goto BEGIN;
...
BEGIN: RenderResult result = template.render(out, context);
if (!result.isDone()) {
  .. save locals..
  state = BEGIN
  return result;
}
```

This optimization was always speculative and the reasoning for why it should apply to `call` commands more than `print` commands is not well supported.  I believe my reasoning was that at the point of a detach there might be 10+ `call` commands on the stack.  While this is true, the same is also true for `print` commands (and to a slightly lesser extend `msg` commands).

The benefit is reduced code size:
[]

And it will make it easier to consolidate the logic between `msg` `print` and `call` commands.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=346110898