Automated [] rollback of changelist 348646430.
*** Reason for rollback ***

triggering a runtime error in some cases that appear to be related to placeholders in plurals

*** Original change description ***

Fix a bug in how TemplateAnalysis works with message placeholders.

Due to a quirk of implementation we would `clone` message placeholders in certain cases prior to generating code for them.  This would cause all `TemplateAnalysis` queries to return `false` since they rely on Node identity.  This is a minor code size issue (and theoretically a minor compiler performance issue).  In the future it will become a correctness issue as we rely on `TemplateAnalysis` in more places.

Code size changes:...

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=348672296