Fix how `msg` nodes are handled in TemplateAnalysis
msg nodes have complex semantics because translators are allowed to reorder and remove placeholders, so we need to be conservative in how they are analyzed.  This bug in the analysis was hidden by another issue in the compiler where message placeholders are always cloned when compiled, which means all the `TemplateAnalysis` queries would fail.  When attempting to fix that issue in cl/348646430 this bug was exposed.

Also add support for visualizing the AccessGraph by implementing a `toString` method that dumps a `dot` formatted string.  I added this to help debug a test failure that ended up being a bug in the test infra, but this seems useful enough to keep.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=349332157