Remove the returntype from `renderInternal` it appears to be unused
I noticed this because of the way that `maybeSkip` logic is generated in normal `templates` issues a `return true;` statement, but this is incompatible with the declared return type.  As far as i can tell this return value is completely unused so we can just drop it rather than trying to line up the types.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=352632989