Fix SEC handling of concatenating attributes.
Currently the entire attribute is suppressed if all conditional values are suppressed, but only if there is a default value. Merge the separate logic branches for with/without default to fix this. Move generation of concatenation logic to a helper function for clarity.

GITHUB_BREAKING_CHANGES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=352833823