Add a stubbing API for soy.
This has every soy template test for the existence of a property in a global object at the beginning of execution. Tests can use this symbol map to swap implementations of the function and then restore them at the end of the test.  A low level stubbing api is supplied which can be used by higher level libraries

This feature is enabled conditionally enabled based `goog.DEBUG`.  If needed we can introduce a new independent `goog.define` for this feature.

There are some basic unit tests to demonstrate behavior.

The code size delta on this is small
[]

I believe the changes are largely due to the soy runtime now being moved around since every .soy.js file now depends on it instead of just the vast majority of them.

This is the third attempt at creating this API, previous attempts were in some ways simpler but much harder to optimize away.  See [] for a description of those other attempts.

[]

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=354959786