Adopt goog.module for the core soy runtime libraries
Move soydata symbols to a legacy aliases class and deprecate!

A previous attempt was to move symbols to a new module and leave the deprecated code in place, however due to some custom code generators using these apis this is hard.  In particular was the method `$$getDeltemplateId` which is an `@idgenerator` and as such we cannot provide a trivial alias for it.  This method should only be used by the soy runtime, but alas we have grown some unfortunate dependencies from weird custom tools.

[]

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=355393486