Add a second call style to jssrc and incrementaldom.
For simple callsites instead of packing up a parameter object, pass each parameter individually as positional parameters.

For example,
```
{call .foo}{param a: XXX/}{param b: YYY/}{/call}
```
is currently compiled to:

```
foo({a:XXX, b: YYY}, opt_ijData)
```

but after this CL would be compiled to

```
foo(opt_ijData, XXX, YYY)
```

NOTE:
  * `opt_ijData` has been moved to the initial position, this object is always passed so to allow for optional `@param`s it is passsed in the initial position
  * an explicit positional parameter is defined for every `@param`.  The ordering of params has now become critical, so `TemplateMetadata` is modified to ensure that all required parameters come before all optional parameters which should in some cases simplify callsites

This should have several benefits:
* code size should be reduced after optimization since code building and destructuring these objects should be reduced
* runtime performance should be improved since there will be fewer object allocations due to our calling convention.
* it should be easier for the jscompiler to detect dead/constant parameters and perform constant propagation on them

Of course the main question is "what counts as a simple callsite?"

There are actually two sets of rules for this, some the apply the the callee and some that apply to the caller.

To be a qualifying callee, you must:
* not have indirect parameters (e.g. no `data="all"` callsites)
* not be an `element` or a `deltemplate`
  * in the future we can support elements simply by changing the implementation
  * in the future we would need new compiler rules to support deltemplates, e.g. if every deltemplate had to implement a defined template signature.

If this is satisfied, then the template may be called using positional style.

To be a qualifying callsite you must
* have a qualifying callee
* cross html context calls in incrementaldom are also not supported, this is a limitation of the current implementation and may also be addressed in the future
* must not be a `data="..."` style callsite
* must not be a dynamic callsite, i.e. `{call $foo}`, this is theoretically fixable but would require some changes to our template literal code generation that are deferred for now.

To support this more flexible call structure we need to generate 2 functions for every qualifying template.  One using the public signature and a second using the positional signature.

for example, given
```soy
{namespace ns}
{template foo}
  {@param a: int}
  {@param b: string}
...
{/template}
```

We currently generate:

```js
ns.foo = function(opt_data, opt_ijData) {
  const a = assertType(opt_data.a, number);
  const b = assertType(opt_data.b, string);
  ...
}
```

we will now generate

```
ns.foo = function(opt_data, opt_ijData) {
  return ns.foo.$$pos(opt_ijData, opt_data.a, opt_data.b);
}
ns.foo$(opt_ijData, a, b) {
  assertType(a, number);
  assertType(b, string);
  ...
}
```

(There are more detailed examples in the golden file tests)

Obviously, when it comes to the definitions, there is no benefit to code size for this transformation.  At best the JsCompiler will inline these functions and generate the original code, at worst this won't happen and we end up with more functions and code size.  However, generally all templates will only have one or the other of these functions called and therefore one of the two will be dead code eliminated, if the `$` function is the one that stays around we will achieve size benefits at the callsites and in the implementation since we no longer will need to unpack parameters.

In theory this should be a completely compatible change since precisely how soy templates call each other is not part of a public API. However, users may rely on the exact names and call signatures of templates in order to support template mocking/stubbing in unit tests.

Code size improvements:
Note, there are too many findings to display them all in [], see them all here []

Summary:
[]

[]

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=355738669