Add a mechanism to save/restore an arbitrary bag of fields to RenderContext
This will be the future basis of our implementation of saving and restoring state to support detach operations in jbcsrc.  For now, there are simply some unit tests for the infrastructure.

There is some awkwardness due to the slow rollout of jdk11 so we need to be compatible with both jdk8 and jdk11 when it comes to dynamically defining classes.

The method `bootstrapSaveState` will be used to save all local variables for a template when we need to suspend execution.  These will be saved into a dynamically generated `StackFrame` subclass.

The method `bootstrapRestoreState` will be used to restore a local variable by dereferencing into that StackFrame object.

This is part of []

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=357056185