Disallow duplicate namespaces.
This is already an error for templates compiled to JS or Python since it will create conflicting namespaces/module names.  This is being banned centrally in order to provide a better error message and also provide the same invariant for server side templates.

This uses an exemption list to downgrade the error to a warning. This will be burned down.

NOTE: if this breaks your project please send a cl adding the failing namespace to the exemption list instead of suggesting a rollback

GITHUB_BREAKING_CHANGES=Soy files will all need to have unique namespaces.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=359070875