Change stubbing to no longer throw ClassNotFoundExceptions to trigger the slowpath.
Instead, teach our linker `ClassLoaderFallbackCallFactory` to detect the test classloader and just always select the slowpath for public templates.

The benefit is that we will soon start defining multiple templates in a single java class so the `ClassNotFoundException` approach will no longer work if we need to load a subset of the templates in a single class.

This is a bit simpler though it does have the downside of putting testonly code in the prod codepath.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=360314682