Use invokedynamic for message defaults.
Currently we generate `static final` fields to hold message defaults and initialize them in constructors.  This approach is good because it ensures that they are constructed exactly once and accessed easily.  However, the downside is that it requires a lot of fields and instructions in `<clinit>` methods.  We can avoid a lot of this by using invokedynamic to construct the defaults and store them directly in the callsite.  This should reduce instructions and fields, and eliminate at least some `<clinit>` methods.

This is especially important because this more efficient encoding unblocks an effort to merge all templates in a single file into a singlet java class.  For files with lots of messages it is possible to exceed the maximum message length for the `<clinit>` method.

[]

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=360330593