Bug-fix: Guard against large allocation when there are zero messages.
The scenario is:

- maskHigh is 0 because 0 msgs.
- bucketMask is -1 >>> 6   <- This is a big number, 67108863
- numBuckets becomes 67108864

This causes the bucketBoundaries array to allocate a 268MB int[]. It's encountered but not fatal in the unit test. I added a case that forces failure in the presence of the unwanted condition.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=362351013