Automated [] rollback of changelist 362351013.
*** Reason for rollback ***

Leads to IndexOutOfBounds in getMsg for lookups in empty bundles.

*** Original change description ***

Bug-fix: Guard against large allocation when there are zero messages.

The scenario is:

- maskHigh is 0 because 0 msgs.
- bucketMask is -1 >>> 6   <- This is a big number, 67108863
- numBuckets becomes 67108864

This causes the bucketBoundaries array to allocate a 268MB int[]. It's encountered but not fatal in the unit test. I added a case that forces failure in the presence of the unwanted condition.

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=362374729