Allow <nobr> tags in dynamic HTML output.
This is needed for sanitizing web content that is still using legacy HTML tags.  This tag is a harmless style tag and is hard to replace with non-dynamic SOY template calls because it can wrap arbirary contents.  It's currently causing visual discrepancies in [] when visualizing web pages because we are failing to write out the internal contents.

I believe that the no| prefix was being used to block the other known HTML tags NOSCRIPT and NOFRAMES and not for NOBR.  I propose we make this more explicit.

GITHUB_BREAKING_CHANGES=N/A

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=374879246