Simplify SoyMsgRawTextPart now that http://openjdk.java.net/jeps/254 has landed.
SoyMsgRawTextPart maintained a memory optimization to store raw text for translations in UTF8 `byte` arrays or UTF16 `char` arrays depending on which was smaller.  This was to avoid the overhead of the internal `String` encoding using UTF16.  As of http://openjdk.java.net/jeps/254 this is no longer the implementation of `String`, instead, internally, `String` has a similar but generally more sophisticated optimization.

In addition to simplifying our implementation this should also improve the performance of rendering since we will no longer need to construct strings or perform utf8 decoding in the rendering path.

PiperOrigin-RevId: 381976794