Avoid reusing variables when handling fallback messages, instead generate fresh ones
This should make it easier for the jscompiler to prove that some of the generated branches are dead without relying on rerunning 'flow' analysis.

The downside if any is that there are slightly _more_ variables, i.o.w. more raw code size, however, this pattern being easier to optimize should make this a non-concern.

PiperOrigin-RevId: 438385588