Update a bunch of opensource versions.
In most cases, this updates to the most recent version.

For proto, this updates to 3.20.0 even though 3.21.2 is available, because 3.20.0 is the latest version supported by https://github.com/bazelbuild/rules_proto/ .

This also switches to using <dependencyManagement> to manage conflicting transitive dependency versions in Maven, as this seems simpler than the previous <exclusions> method.

This removes a Bazel dependency on @maven//:com_google_inject_extensions_guice_multibindings as that has been merged into the main Guice repository.

This only updates rules_java to 4.0.0 (while 5.0.0 is the latest release) because we can't get Bazel version 5+ to work on Kokoro, and rules_java 5 only supports bazel 5.

PiperOrigin-RevId: 462153539