Fix a code size regression in jbcsrc caused by method syntax for protocol buffers
Method call expressions in jbcsrc generate code to `null` check receiver types.  This code ends up being large and in the case of proto calls and redundant (since the code would naturally throw a null pointer exception anyway).

Instead we just outsource responsibility for null checking receiver types to method implementations.  In the case of proto accesses this is trivial, from an audit of other method implementations this also appears to be the case consistently.

This will represent a slight regression in the utility of the error message.

PiperOrigin-RevId: 481668555