When type checking parameters in positional templates always assign to a new local variable.
This is important to work around a type checking [bug](http://b/130651307) in the JsCompiler. Occasionally we resolve param types to a stricter local type to improve type inference, however, that doesn't work in 'positional' templates if the variable is referenced from a closure as is common in idom templates.

This represents a slight regression in gencode size due to the extra variable declarations, but these aliases should be trivial to eliminate during compilation.

PiperOrigin-RevId: 488977002