Avoid null checks in a number of unboxing operations for reference types.
Modify the SoyExpression API to make the decision about whether or not to care about nullishness explicit.

The major benefit here is in looping constructs which want to deal with unboxed lists and do not support null lists.  By simply ignoring nullishness in the unboxing operation we can preserve behavior and avoid some checks.

PiperOrigin-RevId: 545791272