Fix a bug in jssrc where we wouldn't manage variable scopes correctly.
If two local variables had the same name, but had their _use_ and _definition_ interleaved the code generator might get confused.

To fix this we introduce an explicit name scoping mechanism that tracks both soy logical scopes and js logical scopes.  This influences how/when we need to mangle and also reduces the number of names in scope at a time (when exiting scopes we _remove_ the names).

PiperOrigin-RevId: 563740638