Remove the `isLast` parameter to SoyValueProvider.renderAndResolve
When set to `true` this would allow `DetachableContentProvider` to avoid rendering into a secondary buffer.  As such we could reduce the number of temporary buffers.  However, there was only one case where we were able to set this to `true` (`msg` rendering) and that can be handled by simply modifying the SoyValueProvider implementation.

This was added on the theory that we could use static analysis to prove when buffers were unneeded.  However, this is only ever possible for `let` variables that don't escape the templates they are defined in, and in many of those cases we can simply _inline_ them.  Solving this problem with a generic optimization pass is much better than via some runtime logic.  Finally, we never built the static analysis necessary... if we discover more cases, this or a similar mechanism could be recreated.

The truly ideal case is for `param` passing, but absent a global optimization pass that isn't feasible.

PiperOrigin-RevId: 566728881