Use `ConstantDynamic` in jbcsrc to optimize a number of implicit operations.
Because Soys type system is modeled after javascript and we need to implement it in Java, we paper over the differences with a combination of boxing operations (e.g. `IntegerData.valueOf`), numeric conversions (e.g. soy models `int` as java `long` but sometimes we really need a java `int`), and Soy conversions (`coerceToNumber`/`coerceToBoolean`/`coerceToString`).

All of these operations are implicit in the jbcsrc backend and so demand a solution there (rather than a generic solution in, for example, the `OptimizationPass`).  To improve performance we will now model the behavior of Soy constants/literals as Java constants leveraging the `ConstantDynamic` features of jdk11.

Notably,

* When boxing a constant, we promote the boxed value to the constant pool
  - we can only do this for `boolean`/`float`/`int`/`string` values which do not have identity semantics.
  - additionally special case a number of these operations to leverage the standard constants that already exist (e.g. `BooleanData.TRUE`, `StringData.EMPTY`).

* When compiling `let`/`param`/`@state`/`const` values, attempt to model them as Java constants too in order to omit the logic for promoting these values to `static final` fields and `<clinit>` methods.
  - The major benefit here is slightly smaller class files and lazier constant evaluation, now we only calculate the values when accessed rather than when the class is loaded.

From this we expect a performance progression for the following reasons:

 * anything promoted to a constant will eliminate allocations and runtime work constructing constant values
 * constants modeled in this way are meant to be more useful to the JVM for optimization work (similar to a `static final` field)

PiperOrigin-RevId: 568996177