Create an alternate mechanism to `flushLogsAndRender` that avoids being sensitive to evaluation order.
Currently the `flushLogsAndRender` function executes 'synchronously' (as do all externs/plugins), this means that if the current template is rendering into a buffer, the logs may end up in the wrong place.  A trivial demonstration is

```
{call foo}
  {param deferredContent kind="html"}{deferHtml(flushLogsAndRender($someHtml))}{/param}
{/call}
...
{template foo}
  {param deferredContent:html}
  {if $deferredContent}
    {velog ...}<div>{$deferredContent}</div>{/velog}
  {/if}
  ...
```

The conditional test of `deferredContent` by the `if` will trigger the execution of `flushLogsAndRender` which will in turn log in the wrong context.

Per go/soy-evaluation-order evaluation order semantics are _undefined_ and so the `flushLogsAndRender` function is de-facto bugged.  Luckily this was a preexisting widely acknowledged issue :D

The new approach is to
1. build html deferral into the plugin directly using a new soy type `deferred_html`
   - This type is only valid when used as an extern parameter in a trailing position.
2. require that all `deferred_html` functions are only called directly from a print node (like logging functions)
3. require that deferred_html functions can only be called from templates in the same file
   - This restriction is somewhat arbitrary and could be lifted if needed.

Then instead we can continue to use the essential `flushLogsAndRender` strategy of detaching the html from the logs, but arrange to only do it 'lazily' by adding an api to `LoggingAdvisingAppendable` to `append` a `DeferredText` value.

PiperOrigin-RevId: 571142642